LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/historical/coep-navigate-popup-unsafe-inherit.https-expected.txt


FAIL Popup navigating to same-origin with coop unsafe-inherit/coep assert_equals: name expected "" but got "Popup-navigating-to-same-origin-with-coop-unsafe-inherit/coep"
FAIL Popup navigating to same-site with coop unsafe-inherit/coep assert_equals: name expected "" but got "Popup-navigating-to-same-site-with-coop-unsafe-inherit/coep"
FAIL Popup navigating to same-origin with coop unsafe-inherit/no coep assert_equals: name expected "" but got "Popup-navigating-to-same-origin-with-coop-unsafe-inherit/no-coep"
FAIL Popup navigating to same-site with coop unsafe-inherit/no coep assert_equals: name expected "" but got "Popup-navigating-to-same-site-with-coop-unsafe-inherit/no-coep"
